+++ fractal_server/app/routes/api/v2/task_group.py
@@ -4,9 +4,6 @@ from fastapi import APIRouter
 from fastapi import Depends
 from fastapi import HTTPException
 from fastapi import status
-from packaging.version import InvalidVersion
-from packaging.version import Version
-from packaging.version import parse
 from pydantic.types import AwareDatetime
 from sqlmodel import or_
 from sqlmodel import select
@@ -17,13 +14,15 @@ from fractal_server.app.models import LinkUserGroup
 from fractal_server.app.models import UserOAuth
 from fractal_server.app.models.v2 import TaskGroupActivityV2
 from fractal_server.app.models.v2 import TaskGroupV2
-from fractal_server.app.routes.auth import current_user_act_ver_prof
+from fractal_server.app.routes.auth import get_api_guest
+from fractal_server.app.routes.auth import get_api_user
 from fractal_server.app.routes.auth._aux_auth import (
     _get_default_usergroup_id_or_none,
 )
 from fractal_server.app.routes.auth._aux_auth import (
     _verify_user_belongs_to_group,
 )
+from fractal_server.app.routes.aux._versions import _version_sort_key
 from fractal_server.app.schemas.v2 import TaskGroupActivityAction
 from fractal_server.app.schemas.v2 import TaskGroupActivityRead
 from fractal_server.app.schemas.v2 import TaskGroupActivityStatus
@@ -42,26 +41,6 @@ router = APIRouter()
 logger = set_logger(__name__)
 
 
-def _version_sort_key(
-    task_group: TaskGroupV2,
-) -> tuple[int, Version | str | None]:
-    """
-    Returns a tuple used as (reverse) ordering key for TaskGroups in
-    `get_task_group_list`.
-    The TaskGroups with a parsable versions are the first in order,
-    sorted according to the sorting rules of packaging.version.Version.
-    Next in order we have the TaskGroups with non-null non-parsable versions,
-    sorted alphabetically.
-    Last we have the TaskGroups with null version.
-    """
-    if task_group.version is None:
-        return (0, task_group.version)
-    try:
-        return (2, parse(task_group.version))
-    except InvalidVersion:
-        return (1, task_group.version)
-
-
 @router.get("/activity/", response_model=list[TaskGroupActivityRead])
 async def get_task_group_activity_list(
     task_group_activity_id: int | None = None,
@@ -70,7 +49,7 @@ async def get_task_group_activity_list(
     status: TaskGroupActivityStatus | None = None,
     action: TaskGroupActivityAction | None = None,
     timestamp_started_min: AwareDatetime | None = None,
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_guest),
     db: AsyncSession = Depends(get_async_db),
 ) -> list[TaskGroupActivityRead]:
     stm = select(TaskGroupActivityV2).where(
@@ -102,7 +81,7 @@ async def get_task_group_activity_list(
 )
 async def get_task_group_activity(
     task_group_activity_id: int,
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_guest),
     db: AsyncSession = Depends(get_async_db),
 ) -> TaskGroupActivityRead:
     activity = await db.get(TaskGroupActivityV2, task_group_activity_id)
@@ -126,7 +105,7 @@ async def get_task_group_activity(
 
 @router.get("/", response_model=list[tuple[str, list[TaskGroupRead]]])
 async def get_task_group_list(
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_guest),
     db: AsyncSession = Depends(get_async_db),
     only_active: bool = False,
     only_owner: bool = False,
@@ -175,7 +154,7 @@ async def get_task_group_list(
                 await remove_duplicate_task_groups(
                     task_groups=sorted(
                         list(groups),
-                        key=_version_sort_key,
+                        key=lambda group: _version_sort_key(group.version),
                         reverse=True,
                     ),
                     user_id=user.id,
@@ -194,7 +173,7 @@ async def get_task_group_list(
 @router.get("/{task_group_id}/", response_model=TaskGroupRead)
 async def get_task_group(
     task_group_id: int,
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_guest),
     db: AsyncSession = Depends(get_async_db),
 ) -> TaskGroupRead:
     """
@@ -212,7 +191,7 @@ async def get_task_group(
 async def patch_task_group(
     task_group_id: int,
     task_group_update: TaskGroupUpdate,
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_user),
     db: AsyncSession = Depends(get_async_db),
 ) -> TaskGroupRead:
     """