+++ fractal_server/app/routes/api/v2/workflow_import.py
@@ -15,12 +15,12 @@ from fractal_server.app.models.v2 import WorkflowV2
 from fractal_server.app.routes.api.v2._aux_task_group_disambiguation import (
     _disambiguate_task_groups,
 )
-from fractal_server.app.routes.auth import current_user_act_ver_prof
+from fractal_server.app.routes.auth import get_api_user
 from fractal_server.app.routes.auth._aux_auth import (
     _get_default_usergroup_id_or_none,
 )
+from fractal_server.app.routes.aux._versions import _version_sort_key
 from fractal_server.app.schemas.v2 import TaskImport
-from fractal_server.app.schemas.v2 import TaskImportLegacy
 from fractal_server.app.schemas.v2 import WorkflowImport
 from fractal_server.app.schemas.v2 import WorkflowReadWithWarnings
 from fractal_server.app.schemas.v2 import WorkflowTaskCreate
@@ -73,32 +73,6 @@ async def _get_user_accessible_taskgroups(
     return accessible_task_groups
 
 
-async def _get_task_by_source(
-    source: str,
-    task_groups_list: list[TaskGroupV2],
-) -> int | None:
-    """
-    Find task with a given source.
-
-    Args:
-        source: `source` of the task to be imported.
-        task_groups_list: Current list of valid task groups.
-
-    Return:
-        `id` of the matching task, or `None`.
-    """
-    task_id = next(
-        iter(
-            task.id
-            for task_group in task_groups_list
-            for task in task_group.task_list
-            if task.source == source
-        ),
-        None,
-    )
-    return task_id
-
-
 async def _get_task_by_taskimport(
     *,
     task_import: TaskImport,
@@ -141,14 +115,15 @@ async def _get_task_by_taskimport(
         return None
 
     # Determine target `version`
-    # Note that task_import.version cannot be "", due to a validator
     if task_import.version is None:
         logger.debug(
             "[_get_task_by_taskimport] "
             "No version requested, looking for latest."
         )
-        latest_task = max(matching_task_groups, key=lambda tg: tg.version or "")
-        version = latest_task.version
+        version = max(
+            [tg.version for tg in matching_task_groups],
+            key=_version_sort_key,
+        )
         logger.debug(
             f"[_get_task_by_taskimport] Latest version set to {version}."
         )
@@ -213,7 +188,7 @@ async def _get_task_by_taskimport(
 async def import_workflow(
     project_id: int,
     workflow_import: WorkflowImport,
-    user: UserOAuth = Depends(current_user_act_ver_prof),
+    user: UserOAuth = Depends(get_api_user),
     db: AsyncSession = Depends(get_async_db),
 ) -> WorkflowReadWithWarnings:
     """
@@ -246,19 +221,13 @@ async def import_workflow(
     list_task_ids = []
     for wf_task in workflow_import.task_list:
         task_import = wf_task.task
-        if isinstance(task_import, TaskImportLegacy):
-            task_id = await _get_task_by_source(
-                source=task_import.source,
-                task_groups_list=task_group_list,
-            )
-        else:
-            task_id = await _get_task_by_taskimport(
-                task_import=task_import,
-                user_id=user.id,
-                default_group_id=default_group_id,
-                task_groups_list=task_group_list,
-                db=db,
-            )
+        task_id = await _get_task_by_taskimport(
+            task_import=task_import,
+            user_id=user.id,
+            default_group_id=default_group_id,
+            task_groups_list=task_group_list,
+            db=db,
+        )
         if task_id is None:
             raise HTTPException(
                 status_code=status.HTTP_422_UNPROCESSABLE_CONTENT,