+++ tests/fixtures_factories.py
@@ -194,7 +194,7 @@ async def job_factory(db: AsyncSession):
                 dataset.model_dump_json(exclude={"history", "images"})
             ),
             workflow_dump=json.loads(
-                workflow.model_dump_json(exclude={"task_list"})
+                workflow.model_dump_json(exclude={"task_list", "description"})
             ),
             project_dump=json.loads(
                 project.model_dump_json(exclude={"resource_id"})
@@ -233,12 +233,14 @@ async def task_factory(db: AsyncSession):
             "converter_compound",
             "converter_non_parallel",
         ] = "compound",
+        name: str | None = None,
+        version: str | None = None,
         **kwargs,
     ) -> TaskV2:
         args = dict(
             type=type,
-            name=f"task{index}",
-            version=f"{index}",
+            name=name or f"task{index}",
+            version=version,
             command_parallel="cmd_parallel",
             command_non_parallel="cmd_non_parallel",
         )